tests: small updates in folder structure
