+++ tests/integration/deposits/test_delete_deposit.py
